refactor: move hit dice clamp helper to utils
Move the shared hit-dice clamping helper out of HitDiceManager so dialogs and sheet state can import it from the utilities layer.
